requirement.txt
Pillow==7.2.0
websockets==8.1
requests